# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Trento Apartment Service API
  description: |
    API REST per la piattaforma di ricerca e gestione di appartamenti e annunci a Trento.

    Gli endpoint protetti richiedono un token JWT nell'header `Authorization: Bearer <token>`.
    Il token viene rilasciato al login o alla registrazione e ha una durata di 7 giorni (configurabile via `JWT_EXPIRES_IN`).

    **Ruoli utente disponibili:**
    - `utente base` *(default)* — cerca annunci
    - `utente verificato` — cerca annunci (account verificato)
    - `amministratore` — gestisce appartamenti e annunci; può essere privato o agenzia (con partita IVA)
    - `dipendente comune` — operatore del comune; accede alla dashboard statistica TAS Data
    - `inquilino` — inquilino con contratto attivo
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Server di sviluppo locale
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT ottenuto al login o alla registrazione. Payload contiene `sub` (ID utente), `email`, `ruolo`.
  schemas:
    Indirizzo:
      type: object
      required: [via, numero, città, CAP, Stato]
      properties:
        via:
          type: string
          example: Via Roma
        numero:
          type: integer
          example: 10
        città:
          type: string
          example: Trento
        CAP:
          type: string
          example: "38122"
        Stato:
          type: string
          example: Italia
    Posizione:
      type: object
      properties:
        latitudine:
          type: number
          format: double
          example: 46.0748
        longitudine:
          type: number
          format: double
          example: 11.1217
    Camera:
      type: object
      required: [prezzo, mq, tipo]
      properties:
        prezzo:
          type: number
          example: 350
          description: Prezzo mensile in euro
        mq:
          type: number
          example: 14
          description: Superficie della camera in metri quadri
        tipo:
          type: string
          enum: [SINGOLA, DOPPIA]
          example: SINGOLA
        numFinestre:
          type: integer
          default: 1
          example: 2
        disponibile:
          type: boolean
          default: true
          example: true
    Appartamento:
      type: object
      properties:
        _id:
          type: string
          example: 665f1a2b3c4d5e6f7a8b9c0d
        indirizzo:
          $ref: '#/components/schemas/Indirizzo'
        mqTot:
          type: number
          example: 90
          description: Superficie totale in metri quadri (opzionale)
        perStudenti:
          type: boolean
          default: false
          example: false
        numStanze:
          type: integer
          example: 3
        numBagni:
          type: integer
          example: 2
        foto:
          type: array
          items:
            type: string
            format: uri
          example: []
        terrazzo:
          type: boolean
          default: false
          example: true
        lavatrice:
          type: boolean
          default: false
          example: false
        classeEnergetica:
          type: string
          enum: [A4, A3, A2, A1, B, C, D, E, F, G]
          example: B
        posizione:
          $ref: '#/components/schemas/Posizione'
        amministratoreId:
          type: string
          example: 665f1a2b3c4d5e6f7a8b9c01
          description: ID MongoDB dell'amministratore proprietario
        camere:
          type: array
          items:
            $ref: '#/components/schemas/Camera'
...
    Paginazione:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 5
        totalItems:
          type: integer
          example: 48
        itemsPerPage:
          type: integer
          example: 10

    # -- US16: strutture dati per la dashboard statistica --
...
    ErroreGenerico:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Errore interno del server
    ErroreAutenticazione:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Token non fornito. Usa Authorization: Bearer <token>"
    ErroreAutorizzazione:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Accesso non autorizzato

# ---------------------------------------------------------------------------
# PATH
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/appartamenti:
    get:
      tags: [Appartamenti]
      summary: Lista appartamenti
      description: |
        Restituisce la lista paginata degli appartamenti. Accessibile senza autenticazione.
        Con `proprietario=me` e un token JWT valido restituisce solo gli appartamenti
        dell'amministratore autenticato (l'`amministratoreId` è ricavato dal token).
      operationId: getAppartamenti
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Numero di pagina
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Elementi per pagina
        - name: perStudenti
          in: query
          schema:
            type: boolean
          description: Filtra appartamenti adatti agli studenti
        - name: città
          in: query
          schema:
            type: string
          description: Filtra per città (case-insensitive, ricerca parziale)
        - name: proprietario
          in: query
          schema:
            type: string
            enum: [me]
          description: Se valorizzato a `me`, restituisce solo gli appartamenti dell'amministratore autenticato (richiede token JWT)
      responses:
        "200":
          description: Lista paginata degli appartamenti
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Appartamenti recuperati
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Appartamento'
                  pagination:
                    $ref: '#/components/schemas/Paginazione'
        "401":
          description: Richiesto `proprietario=me` senza token JWT valido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutenticazione'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'

    post:
      tags: [Appartamenti]
      summary: Crea appartamento
      description: |
        Crea un nuovo appartamento. Richiede autenticazione JWT.
        L'`amministratoreId` viene estratto automaticamente dal token JWT (`sub`).
        La `posizione` (lat/lon) viene calcolata automaticamente dall'indirizzo
        tramite geocoding (Nominatim/OpenStreetMap) se non fornita dal client.
      operationId: creaAppartamento
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [indirizzo, numStanze, numBagni]
              properties:
                indirizzo:
                  $ref: '#/components/schemas/Indirizzo'
                mqTot:
                  type: number
                  example: 75
                perStudenti:
                  type: boolean
                  default: false
                numStanze:
                  type: integer
                  example: 3
                numBagni:
                  type: integer
                  example: 1
                foto:
                  type: array
                  items:
                    type: string
                    format: uri
                terrazzo:
                  type: boolean
                  default: false
                lavatrice:
                  type: boolean
                  default: false
                classeEnergetica:
                  type: string
                  enum: [A4, A3, A2, A1, B, C, D, E, F, G]
                posizione:
                  $ref: '#/components/schemas/Posizione'
                camere:
                  type: array
                  items:
                    $ref: '#/components/schemas/Camera'
      responses:
        "201":
          description: Appartamento creato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Appartamento registrato con successo
                  data:
                    $ref: '#/components/schemas/Appartamento'
        "400":
          description: Errore di validazione del body
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                  errors:
                    type: array
                    items:
                      type: string
        "401":
          description: Token JWT assente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutenticazione'
        "403":
          description: Token JWT non valido o scaduto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutorizzazione'
...
  /api/v1/appartamenti/{id}:
    get:
      tags: [Appartamenti]
      summary: Dettaglio appartamento
      description: Restituisce i dati completi di un singolo appartamento. Non richiede autenticazione.
      operationId: getAppartamentoById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId dell'appartamento
      responses:
        "200":
          description: Dati dell'appartamento
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Appartamento recuperato
                  data:
                    $ref: '#/components/schemas/Appartamento'
        "404":
          description: Appartamento non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'

    put:
      tags: [Appartamenti]
      summary: Aggiorna appartamento
      description: |
        Aggiorna i campi di un appartamento esistente.
        Richiede autenticazione JWT e che l'utente sia il proprietario (`amministratoreId`).
        Tutti i campi del body sono opzionali; i campi non inclusi non vengono modificati.
        Se l'`indirizzo` viene modificato, la `posizione` viene ricalcolata via geocoding.
      operationId: aggiornaAppartamento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId dell'appartamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                indirizzo:
                  $ref: '#/components/schemas/Indirizzo'
                mqTot:
                  type: number
                perStudenti:
                  type: boolean
                numStanze:
                  type: integer
                numBagni:
                  type: integer
                terrazzo:
                  type: boolean
                lavatrice:
                  type: boolean
                classeEnergetica:
                  type: string
                  enum: [A4, A3, A2, A1, B, C, D, E, F, G]
                camere:
                  type: array
                  items:
                    $ref: '#/components/schemas/Camera'
      responses:
        "200":
          description: Appartamento aggiornato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Appartamento aggiornato con successo
                  data:
                    $ref: '#/components/schemas/Appartamento'
        "400":
          description: Errori di validazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "401":
          description: Token assente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutenticazione'
        "403":
          description: Token non valido o utente non proprietario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutorizzazione'
        "404":
          description: Appartamento non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'

    delete:
      tags: [Appartamenti]
      summary: Elimina appartamento
      description: |
        Elimina un appartamento. Richiede autenticazione JWT e che l'utente sia il proprietario.
        Se esistono annunci collegati, la risposta include il campo `warnings` con il conteggio degli annunci orfani.
      operationId: eliminaAppartamento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId dell'appartamento
      responses:
        "200":
          description: Appartamento eliminato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Appartamento eliminato con successo
                  data:
                    $ref: '#/components/schemas/Appartamento'
                  warnings:
                    type: string
                    nullable: true
                    example: Erano presenti 2 annunci associati
        "401":
          description: Token assente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutenticazione'
        "403":
          description: Token non valido o utente non proprietario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutorizzazione'
        "404":
          description: Appartamento non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /api/v1/appartamenti/{id}/contatto-admin:
    get:
      tags: [Appartamenti]
      summary: Contatto WhatsApp dell'amministratore
      description: >-
        Restituisce nome e link WhatsApp dell'amministratore dell'appartamento.
        Riservato agli utenti autenticati (registrati): alimenta il pulsante
        "Contatta su WhatsApp" nella pagina dell'annuncio.
      operationId: getContattoAdmin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID dell'appartamento
      responses:
        "200":
          description: Contatto recuperato
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      nome:
                        type: string
                        example: Luca
                      cognome:
                        type: string
                        example: Bianchi
                      telefono:
                        type: string
                        nullable: true
                        example: "+39 351 1234567"
                      linkWhatsApp:
                        type: string
                        nullable: true
                        description: Link click-to-chat (wa.me) gia' pronto; null se l'admin non ha un numero valido
                        example: "https://wa.me/393511234567?text=Salve%2C%20sono%20interessato..."
        "401":
          description: Token assente o non valido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutenticazione'
        "404":
          description: Appartamento o amministratore non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /api/v1/appartamenti/{appartamentoId}/associa-inquilino:
    post:
      tags: [Appartamenti]
      summary: Associa un inquilino a un appartamento (contratto)
      description: >-
        Crea il legame amministratore->inquilino tramite contratto. Eseguibile
        solo dall'amministratore proprietario. L'utente, indicato per email,
        dev'essere verificato/inquilino e senza contratti attivi. Se l'appartamento
        ha gia' un contratto attivo l'utente viene aggiunto come coinquilino,
        altrimenti se ne crea uno nuovo (servono date, canone e tipo). A fine
        operazione l'utente assume il ruolo `inquilino`.
      operationId: associaInquilino
      security:
        - bearerAuth: []
      parameters:
        - name: appartamentoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: greta.delai@example.com
                dataInizio:
                  type: string
                  format: date
                  description: Obbligatoria solo se si crea un nuovo contratto
                  example: "2026-01-01"
                dataFine:
                  type: string
                  format: date
                  description: Obbligatoria solo se si crea un nuovo contratto
                  example: "2026-12-31"
                canoneMensile:
                  type: number
                  description: Obbligatorio solo se si crea un nuovo contratto
                  example: 850
                tipoContratto:
                  type: string
                  description: Obbligatorio solo se si crea un nuovo contratto
                  example: Residenziale
      responses:
        "200":
          description: Inquilino associato (contratto creato o aggiornato)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Greta Delai associato all'appartamento
                  data:
                    $ref: '#/components/schemas/Contratto'
        "400":
          description: Email mancante, utente non verificato o dati contratto incompleti
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "403":
          description: Non autenticato come amministratore proprietario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutorizzazione'
        "404":
          description: Nessun utente registrato con questa email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "409":
          description: L'utente ha gia' un contratto attivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'

  # ── Utenti ────────────────────────────────────────────────────────────────
...